docs(api): field level documentation - PART 4
-Documented the fields for even more entities

Changes:
-added more fields for Win32DiskDrive

-Updated the some numeric fields (with uint64 return type in MS docs) in some classes to have BigInteger instead of Long as its return type

- Corresponding tests were updated

+++ src/main/java/io/github/eggy03/ferrumx/windows/entity/memory/Win32PhysicalMemory.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Value;
 import org.jetbrains.annotations.Nullable;
 
+import java.math.BigInteger;
+
 /**
  * Immutable representation of a RAM module on a Windows system.
  * <p>
@@ -129,7 +131,7 @@ public class Win32PhysicalMemory {
      */
     @SerializedName("Capacity")
     @Nullable
-    Long capacity;
+    BigInteger capacity;
 
     /**
      * Data width of the physical memory—in bits.
