+++ src/main/java/io/github/eggy03/ferrumx/windows/entity/storage/Win32DiskPartition.java
@@ -8,8 +8,10 @@ import lombok.Getter;
 import lombok.Value;
 import org.jetbrains.annotations.Nullable;
 
+import java.math.BigInteger;
+
 /**
- * Immutable representation of a partition in a <b>physical</b> disk on a Windows system.
+ * Immutable representation of a <b>Partition</b> in a <b>Physical</b> disk on a Windows system.
  * <p>
  * Fields correspond to properties retrieved from the {@code Win32_DiskPartition} WMI class.
  * </p>
@@ -52,56 +54,113 @@ import org.jetbrains.annotations.Nullable;
 @Builder(toBuilder = true)
 public class Win32DiskPartition {
 
+    /**
+     * Unique identifier of the disk drive and partition within the system.
+     */
     @SerializedName("DeviceID")
     @Nullable
     String deviceId;
 
+    /**
+     * Label by which the partition is known.
+     */
     @SerializedName("Name")
     @Nullable
     String name;
 
+    /**
+     * Description of the partition object.
+     */
     @SerializedName("Description")
     @Nullable
     String description;
 
+    /**
+     * Size in bytes of the blocks that form this partition.
+     */
     @SerializedName("BlockSize")
     @Nullable
-    Long blockSize;
+    BigInteger blockSize;
 
+    /**
+     * Total number of consecutive blocks that form this partition.
+     * The total size of the partition can be calculated by multiplying this value by {@link #blockSize}.
+     */
     @SerializedName("NumberOfBlocks")
     @Nullable
-    Long numberOfBlocks;
+    BigInteger numberOfBlocks;
 
+    /**
+     * Indicates whether the computer can be booted from this partition.
+     */
     @Getter(AccessLevel.NONE)
     @SerializedName("Bootable")
     @Nullable
     Boolean bootable;
     public @Nullable Boolean isBootable() {return bootable;}
 
+    /**
+     * Indicates whether this is the primary partition on the disk.
+     */
     @Getter(AccessLevel.NONE)
     @SerializedName("PrimaryPartition")
     @Nullable
     Boolean primaryPartition;
     public @Nullable Boolean isPrimaryPartition() {return primaryPartition;}
 
+    /**
+     * Indicates whether this is the active (boot) partition used by the operating system when booting.
+     */
     @Getter(AccessLevel.NONE)
     @SerializedName("BootPartition")
     @Nullable
     Boolean bootPartition;
     public @Nullable Boolean isBootPartition() {return bootPartition;}
 
+    /**
+     * Index number of the physical disk that contains this partition.
+     */
     @SerializedName("DiskIndex")
     @Nullable
     Long diskIndex;
 
+    /**
+     * Total size of the partition in bytes.
+     */
     @SerializedName("Size")
     @Nullable
-    Long size;
+    BigInteger size;
 
+    /**
+     * Type of the partition
+     * <p>Possible Values (Non-exhaustive, will be updated when new values are found):</p>
+     * <ul>
+     *     <li>Unused</li>
+     *     <li>12-bit FAT</li>
+     *     <li>Xenix Type 1</li>
+     *     <li>Xenix Type 2</li>
+     *     <li>16-bit FAT</li>
+     *     <li>Extended Partition</li>
+     *     <li>MS-DOS V4 Huge</li>
+     *     <li>Installable File System</li>
+     *     <li>PowerPC Reference Platform</li>
+     *     <li>UNIX</li>
+     *     <li>NTFS</li>
+     *     <li>ReFS</li>
+     *     <li>Win95 w/Extended Int 13</li>
+     *     <li>Extended w/Extended Int 13</li>
+     *     <li>Logical Disk Manager</li>
+     *     <li>Unknown</li>
+     * </ul>
+     */
     @SerializedName("Type")
     @Nullable
     String type;
 
+    /**
+     * Prints the entity in a JSON pretty-print format
+     * @return the {@link String} value of the object in JSON pretty-print format
+     */
     @Override
     public String toString() {
         return new GsonBuilder()
